Moves serializer and model out from api, preview now working
